# Batch fingerprint encodings