--- a/src/command_line/indexing/bulk_index.rs
+++ b/src/command_line/indexing/bulk_index.rs
@@ -1,14 +1,8 @@
 use crate::command_line::{indexing::split_path, prelude::*};
 use crate::indexing::index_manager::IndexManager;
-use crate::search::compound_processing::process_cpd;
-use crate::search::similarity_search::encode_fingerprints;
 use rayon::prelude::*;
-use std::{collections::HashMap, fs::File, io::BufRead, io::BufReader, ops::Deref};
-use bitvec::prelude::BitVec;
-use rdkit::{Fingerprint, ROMol};
+use std::{fs::File, io::BufRead, io::BufReader, ops::Deref};
 use serde_json::Value;
-use tantivy::Document;
-use tantivy::schema::{Field, Schema};
 
 pub const NAME: &str = "bulk-index";
 
@@ -48,48 +42,67 @@ pub fn action(matches: &ArgMatches) -> eyre::Result<()> {
     let file = File::open(json_path)?;
     let reader = BufReader::new(file);
     let chunksize = 1000;
-    let mut record_vec = Vec::with_capacity(chunksize);
+    let mut compound_vec = Vec::with_capacity(chunksize);
 
     for result_line in reader.lines() {
         let line = result_line?;
-        let record = serde_json::from_str(&line)?;
+        let record = &serde_json::from_str(&line)?;
+        let smiles_and_extra_data = get_smiles_and_extra_data(record)?;
 
-        record_vec.push(record);
-        if record_vec.len() == chunksize {
-            match batch_doc_creation(&mut record_vec, &schema) {
-                Err(e) => log::warn!("Doc creation batch failed: {e}"),
+        compound_vec.push(smiles_and_extra_data);
+        if compound_vec.len() == chunksize {
+            match batch_doc_creation(&compound_vec, &schema) {
+                Err(e) => log::warn!("Failed batched doc creation: {e}"),
                 Ok(doc_batch) => {
                     let _ = doc_batch
                         .into_par_iter()
                         .map(|doc| {
-                            match writer.add_document(doc) {
-                                Ok(_) => (),
-                                Err(_) => {
-                                    log::warn!("Failed doc creation");
+                            match doc {
+                                Ok(doc) => {
+                                    match writer.add_document(doc) {
+                                        Ok(_) => (),
+                                        Err(_) => {
+                                            log::warn!("Failed doc creation: Could not add document");
+                                        }
+                                    }
+                                },
+                                Err(e) => {
+                                    log::warn!("Failed doc creation: {e}");
                                 }
                             }
                         }).collect::<Vec<()>>();
                 }
             }
+
+            compound_vec.clear();
         }
     }
 
-    if !record_vec.is_empty() {
-        match batch_doc_creation(&mut record_vec, &schema) {
+    if !compound_vec.is_empty() {
+        match batch_doc_creation(&compound_vec, &schema) {
             Err(e) => log::warn!("Doc creation batch failed: {e}"),
             Ok(doc_batch) => {
                 let _ = doc_batch
                     .into_par_iter()
                     .map(|doc| {
-                        match writer.add_document(doc) {
-                            Ok(_) => (),
-                            Err(_) => {
-                                log::warn!("Failed doc creation");
+                        match doc {
+                            Ok(doc) => {
+                                match writer.add_document(doc) {
+                                    Ok(_) => (),
+                                    Err(_) => {
+                                        log::warn!("Failed doc creation: Could not add document");
+                                    }
+                                }
+                            },
+                            Err(e) => {
+                                log::warn!("Failed doc creation: {e}");
                             }
                         }
                     }).collect::<Vec<()>>();
             }
         }
+
+        compound_vec.clear();
     }
 
     let _ = writer.commit();
@@ -97,84 +110,13 @@ pub fn action(matches: &ArgMatches) -> eyre::Result<()> {
     Ok(())
 }
 
-fn batch_doc_creation(
-    record_vec: &mut Vec<Value>,
-    schema: &Schema,
-) -> eyre::Result<Vec<impl Document>> {
-    let smiles_field = schema.get_field("smiles")?;
-    let pattern_fingerprint_field = schema.get_field("pattern_fingerprint")?;
-    let morgan_fingerprint_field = schema.get_field("morgan_fingerprint")?;
-    let extra_data_field = schema.get_field("extra_data")?;
-    let other_descriptors_field = schema.get_field("other_descriptors")?;
-    let descriptor_fields = KNOWN_DESCRIPTORS
-        .iter()
-        .map(|kd| (*kd, schema.get_field(kd).unwrap()))
-        .collect::<HashMap<&str, Field>>();
-
-    let mol_attributes = record_vec
-        .clone()
-        .into_par_iter()
-        .filter_map(|r| {
-            let extra_data = r.get("extra_data").cloned();
-            if let Some(smiles) = r.get("smiles") {
-                if let Some(smiles) = smiles.as_str() {
-                    match process_cpd(smiles, false) {
-                        Ok(attributes) => {
-                            Some((attributes.0, extra_data, attributes.1, attributes.2))
-                        },
-                        Err(e) => {
-                            log::warn!("Failed compound processing for smiles '{}': {}", smiles, e);
-                            None
-                        }
-                    }
-                } else {
-                    log::warn!("Failed to convert smiles to str");
-                    None
-                }
-            } else {
-                log::warn!("Failed to extract smiles");
-                None
-            }
-        }).collect::<Vec<(ROMol, Option<Value>, Fingerprint, HashMap<String, f64>)>>();
-
-    record_vec.clear();
-
-    let mut morgan_fingerprints: Vec<Fingerprint> = Vec::with_capacity(mol_attributes.len());
-    let mut morgan_bitvecs: Vec<BitVec<u8>> = Vec::with_capacity(mol_attributes.len());
-    for attributes in mol_attributes.clone() {
-        let morgan_fp = attributes.0.morgan_fingerprint();
-        morgan_fingerprints.push(morgan_fp.clone());
-        morgan_bitvecs.push(morgan_fp.0);
-    }
-
-    let similarity_clusters = encode_fingerprints(&morgan_bitvecs, true)
-        .map_err(|e| eyre::eyre!("Failed batched similarity cluster assignment: {e}"))?;
-
-    let docs = (0..mol_attributes.len())
-        .into_iter()
-        .filter_map(|i| {
-            let attributes = &mol_attributes[i];
-            match create_tantivy_doc(
-                &attributes.0,
-                &attributes.1,
-                &attributes.2,
-                &morgan_fingerprints[i],
-                &attributes.3,
-                similarity_clusters[i],
-                smiles_field,
-                pattern_fingerprint_field,
-                morgan_fingerprint_field,
-                &descriptor_fields,
-                extra_data_field,
-                other_descriptors_field,
-            ) {
-                Ok(doc) => Some(doc),
-                Err(_) => {
-                    log::warn!("Failed doc creation");
-                    None
-                },
-            }
-        }).collect::<Vec<_>>();
-
-    Ok(docs)
+fn get_smiles_and_extra_data(record: &Value) -> eyre::Result<(String, Option<Value>)>{
+    let smiles = record
+        .get("smiles")
+        .ok_or(eyre::eyre!("Failed to extract smiles"))?
+        .as_str()
+        .ok_or(eyre::eyre!("Failed to parse smiles"))?
+        .to_string();
+    let extra_data = record.get("extra_data").cloned();
+    Ok((smiles, extra_data))
 }