… Cargo.lock
… Cargo.toml
… src/command_line/search/similarity_search.rs
… src/indexing/mod.rs
… src/rest_api/api/indexing/bulk_index.rs
… src/rest_api/api/search/similarity_search.rs
… src/search/identity_search.rs
… src/search/scaffold_search.rs
… src/search/similarity_search.rs
… src/search/structure_search.rs
… tests/api_tests.rs