Resolves [#129](https://github.com/rdkit-rs/cheminee/issues/129). Previously, the tensorflow similarity model was being built for every compound that was indexed (in an effort to avoid null pointer issues in tensorflow rust). Now we are batching the fingerprint encoding step (i.e. relies on tensorflow rust), which significantly cuts down the number of times we rebuild the tensorflow similarity model. Importantly, this still allows us to avoid null pointer issues.

Local tests seem to indicate a 536% improvement in indexing speed for the `index-sdf` CLI endpoint.

**Benchmarks**

Before PR:
```
cargo run --release -- create-index -i tmp/cheminee/test_index_20250107 -n descriptor_v1
time TF_CPP_MIN_LOG_LEVEL=3 cargo run --release -- index-sdf -i tmp/cheminee/test_index_20250107 -s tmp/sdfs/Compound_000000001_000500000.sdf.gz -l 50000

3125.72s user 426.14s system 1103% cpu 5:21.95 total
```

After PR:
```
cargo run --release -- create-index -i tmp/cheminee/test_index_20250107 -n descriptor_v1
time TF_CPP_MIN_LOG_LEVEL=3 cargo run --release -- index-sdf -i tmp/cheminee/test_index_20250107 -s tmp/sdfs/Compound_000000001_000500000.sdf.gz -l 50000

583.29s user 143.53s system 740% cpu 1:38.14 total
```